% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive}
\alias{archive}
\title{Archive}
\usage{
archive(fileName = NULL, currentDir = NULL, archiveDir = NULL,
  verbose = FALSE)
}
\arguments{
\item{fileName}{Character vector. The name of the file that must be archived}

\item{currentDir}{Character vector. The current directory of the file}

\item{archiveDir}{Character vector. The directory that the file must be archived in}

\item{verbose}{Logical. Displays function messages when TRUE}
}
\description{
Function is used to archive an R data file
}

