% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binconf_uur.R
\name{binconf_uur}
\alias{binconf_uur}
\title{Binary Configuration Hour}
\usage{
binconf_uur(x = stowe.df, datumvar = "datum", vuurvar = "vuur",
  vuurnaam = "fire", plot = c(NULL, "bar", "line"), stoor = FALSE,
  xlsx = TRUE, stoordir = NULL, fn = "binconf.uur.pdf",
  ttl = "Proportion of fires active\\n by hour of day (with 95\% CI)",
  addttl = NULL, addfn = NULL, ylm = c(0, 1), verbose = TRUE,
  ylable = "Proportion of fires active", uurvar = "hour",
  maandvar = "month", jaarvar = "year")
}
\arguments{
\item{x}{Data frame that contains fire data}

\item{datumvar}{Character vector. Name of the column containing dates}

\item{vuurvar}{Character vector. Name of the column containing fires being active or not}

\item{vuurnaam}{Character vector. Is the fire active? Either "fire" or "yes".}

\item{plot}{Character vector. Plot nothing (NULL), barplot (bar), or lineplot (line)}

\item{stoor}{Logical. Store the data?}

\item{xlsx}{Logical. Print out xlsx or not.}

\item{stoordir}{Character vector. The directory for storage to take place in}

\item{fn}{Character vector. Filename of stored data}

\item{ttl}{Character vector. Title of the generated graph}

\item{addttl}{Character vector. Add a title}

\item{addfn}{Character vector. Add a filename}

\item{ylm}{Numeric vector. Contains y axis limits}

\item{verbose}{Logical. Displays messages}

\item{ylable}{Character vector. The y-axis lable}

\item{uurvar}{Character vector. Name of the hours variable. If null the function will
extract the hour of day from the data contained in the date column (recommended)}

\item{maandvar}{Character vector. Name of the month variable}

\item{jaarvar}{Character vector. Name of the year variable}
}
\description{
Function receives a data frame and creates a graph based on the
proportion of fires active by hour. Also gives xlsx output of prop, upper and lower. Makes use
of Monte Carlo Simulation
}

