% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_fires.R
\name{count.fires.per.day.list}
\alias{count.fires.per.day.list}
\title{Count Fires Per Day List}
\usage{
count.fires.per.day.list(ldata = NULL, returnAsSingleDataFrame = FALSE, ...)
}
\arguments{
\item{ldata}{List containing fire information}

\item{returnAsSingleDataFrame}{Logical. Returns the results as is if TRUE}

\item{...}{Arguments to be passed to/from other methods}
}
\description{
Uses the count.fire.per.day function while receiving a list. Function returns raw counts instead
of summarised ones
}

