% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maak.vuur.R
\name{maak.vuur}
\alias{maak.vuur}
\title{Maak Vuur}
\usage{
maak.vuur(df, aggr = FALSE, drops = c("fire", "E"), datum.naam = "date",
  skalleer = TRUE, stoofnaam = "c", verbose = FALSE, debug = FALSE,
  tpp = "p", qp = FALSE)
}
\arguments{
\item{df}{Data frame. Contains fire and temperature data, normally the iButton data}

\item{aggr}{Logical. Initialises date aggregation when TRUE}

\item{drops}{Character vector. Names of the columns to be dropped}

\item{datum.naam}{Character vector. Name of the column containing dates}

\item{skalleer}{Logical. If TRUE then data are scaled to unit variance}

\item{stoofnaam}{Character vector. Name of the column containing chimney temperatures}

\item{verbose}{Logical. Display function messages}

\item{debug}{Logical. Assign values to global environment for debugging purposes}

\item{tpp}{Character vector. The plot type}

\item{qp}{Logical. Make qplot or not}
}
\description{
Main function that receives a data frame containing fire and temperature data and produces multiple plots
that describes the data. Also makes use of the vuur.diagnoseer.plot function.
}
\details{
Use this function in the form df = maak.vuur(df)
or res = lapply(kortLYS, maak.vuur, aggr = TRUE, verbose=T, qp = F)
}

