% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.col.R
\name{move.col}
\alias{move.col}
\title{Move Column}
\usage{
move.col(df = NULL, colName = NULL, colIdx = NULL)
}
\arguments{
\item{df}{Data frame}

\item{colName}{Character vector. Name of the column to be moved}

\item{colIdx}{Numeric. Index to which the column should be moved}
}
\description{
Function to move any given column in a given data frame to any desired index in the data frame
}

