% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.col.R
\name{move.cols}
\alias{move.cols}
\title{Move Columns}
\usage{
move.cols(df = NULL, colNames = NULL, colIdxx = NULL, verbose = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{colNames}{Character vector. Names of the columns to be moved}

\item{colIdxx}{Numeric. Indices to which the columns should be moved}

\item{verbose}{Logical. Displays function messages whn TRUE}
}
\description{
Function to move any number of columns in a given data frame to any desired indices at once
}

