% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maak.vuur.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Arguments to be passed to/from other methods}

\item{plotlist}{A list of ggplot objects}

\item{file}{Filename (Not present in this function)}

\item{cols}{Numeric. Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored}
}
\description{
Plot multiple graphs on one page
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\references{
Winston Chang: http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
Multiple plot function
}

