% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maak.vuur.R
\name{pl}
\alias{pl}
\title{Strip function}
\usage{
pl(x = 1:nrow(dpca$ind$contrib), y = dpca$call$X[, sstf], tp = "p",
  sdif = "sdif", sdifval = 3, pcadf = cluscont, cl, clII, mn, nm, nmII,
  envr = zz, qplot = TRUE, verbose = TRUE, naam = "naam")
}
\description{
Strip function
}

