% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.col.R
\name{remove.col}
\alias{remove.col}
\title{Remove Column}
\usage{
remove.col(df = NULL, colName = NULL, colIdx = NULL)
}
\arguments{
\item{df}{Data frame}

\item{colName}{Character vector. Name of the column to be removed. If null the function uses the idx
and vice versa}

\item{colIdx}{Numeric. Index of the column to be removed}
}
\description{
Function to remove any given column from a given data frame
}

