% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.col.R
\name{remove.cols}
\alias{remove.cols}
\title{Remove Columns}
\usage{
remove.cols(df = NULL, colNames = NULL, colIdxx = NULL)
}
\arguments{
\item{df}{Data frame}

\item{colNames}{Character vector. Name of the columns to be removed. If null the function uses the idx
and vice versa}

\item{colIdxx}{Numeric. Indices of the columns to be removed}
}
\description{
Function to remove any number of columns from a given data frame at once
}

