% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skoonmaakvlamme.R
\name{skoonmaakvlamme}
\alias{skoonmaakvlamme}
\title{Skoonmaak Vlamme}
\usage{
skoonmaakvlamme(df = NULL, fireField = "fire", fireLabel = "fire",
  noFireLabel = "no", chimneyField = "C", dateField = "date",
  timeStep = 20, verbose = FALSE)
}
\arguments{
\item{df}{Data frame. Contains fire data}

\item{fireField}{Character vector. Name of the fire column}

\item{fireLabel}{Character vector. Name indicating active fire}

\item{noFireLabel}{Character vector. Name indicating inactive fire}

\item{chimneyField}{Character vector. Column containing chimney values}

\item{dateField}{Character vector. Name of the date column}

\item{timeStep}{Numeric. The number of minutes between
two consecutive records in the received df}

\item{verbose}{Logical. Display function messages?}
}
\description{
Function is used to remove the cleaning flame temperatures at the beginning or end of
certain fires. It does this by testing if there are any unreasonably high temperatures
in the first or last 10 minutes of a fire episode
}
\details{
THIS FUNCTION ASSUMES THAT THE ARGUMENT RECEIVED BY DF ARE SORTED
ACCORDING TO DATE/TIME - ASCENDINGLY
}

