% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_fires.R
\name{summarise.fires.per.day.list}
\alias{summarise.fires.per.day.list}
\title{Summarise Fires Per Day List}
\usage{
summarise.fires.per.day.list(lys, groupvar = "yr", ...)
}
\arguments{
\item{lys}{List containing fire information}

\item{groupvar}{Character vector. Grouping variable that defaults to "year"}

\item{...}{Arguments to be passed to/from other methods}
}
\description{
Uses the count.fire.per.day function while receiving a list. Function returns summarised counts
}

