% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_fire_classifier.R
\name{svm_fire_classifier}
\alias{svm_fire_classifier}
\title{SVM(Support Vector Machine) Fire Classifier}
\usage{
svm_fire_classifier()
}
\description{
Script for building and training the SVM ibutton fire classification model.
}

