% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maak.vuur.R
\name{vuur.diagnoseer.plot}
\alias{vuur.diagnoseer.plot}
\title{Vuur Diagnoseer Plot}
\usage{
vuur.diagnoseer.plot(dpca = get("dpca", envir = xx), code = zz$stt,
  sstf = "c", debug = TRUE, grafdir = "Grafieke/", naam = get("naam",
  environment(vuur.diagnoseer.plot)))
}
\arguments{
\item{dpca}{List such as the PCA generated from maak.vuur}

\item{code}{Information regaring the zz environment generated in maak.vuur}

\item{sstf}{Character vector. Name of the column containing stove temperatures}

\item{debug}{Logical. Assignments for debugging process}

\item{grafdir}{Character vector. Directory to which graphs will be saved as .pdf}

\item{naam}{Character vector. Header for the temperature graphs}
}
\description{
Function used to create multiple plots of the PCA output from function maak.vuur.
}

