% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maak.vuur.R
\name{vuur.publikasie.plot}
\alias{vuur.publikasie.plot}
\title{Vuur Publikasie Plot}
\usage{
vuur.publikasie.plot(x, vuurnaam = "vuur", sdif = "sdif", sstf = "c",
  naam = "", grafdir = "Grafieke/", verbose = FALSE, save = TRUE,
  DEBUG = FALSE)
}
\arguments{
\item{x}{Data frame. Fire data such as the data output from the maak.vuur function}

\item{vuurnaam}{Character vector. Name of the column indicating whether fire is active or not}

\item{sdif}{Character vector. Name of the column containing the difference in stove and wall temperatures}

\item{sstf}{Character vector. Name of the column containing chimney temperatures}

\item{naam}{Character vector. Name that will appear in the header of the graphs}

\item{grafdir}{Character vector. Directory that the graph will be saved in as a .pdf. The default
directory will be inside a Grafieke folder in your working directory}

\item{verbose}{Logical. Display function messages}

\item{save}{Logical. Saves the graphs in pdf format inside grafdir when TRUE}

\item{DEBUG}{Logical. Assigns values for debugging purposes}
}
\description{
Creates multiple plots relating to chimney, wall and fire temperatures as well as plotting whether
fires were made or not throughout the day
}

