% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wees.vure.R
\name{wees.vure.dood}
\alias{wees.vure.dood}
\title{Wees vure Dood}
\usage{
wees.vure.dood(df = NULL, fireField = "fire", fireLabel = "fire",
  noFireLabel = "no", minFireLength = 2)
}
\arguments{
\item{df}{Data frame. Fire data}

\item{fireField}{Character vector. Name of the fire column}

\item{fireLabel}{Character vector. Name of the label indicating active fire}

\item{noFireLabel}{Character vector. Name of the label indicating non-active fire}

\item{minFireLength}{Numeric. The minimum duration allowed for a fire episode - given as
the number of records/time steps across which it should span}
}
\description{
Function analyses fire episodes and isolates those that are seen as too short
(specified by minFireLength numeric), whereby it removes them
}

