% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wees.vure.R
\name{wees.vure}
\alias{wees.vure}
\title{Wees Vure}
\usage{
wees.vure(x, vuurvar = "fire", verbose = FALSE, wees = 2)
}
\arguments{
\item{x}{Data frame. Fire data}

\item{vuurvar}{Character vector. Name of the fire column}

\item{verbose}{Logical. Display function messages}

\item{wees}{Numeric. The minimum duration allowed for a fire episode - given as the
number of records/time steps across which it should span}
}
\description{
Function analyses fire episodes and isolates those that are seen as too short
(specified by "wees" numeric), whereby it replaces the episodes with longer ones
}

