% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomVuur.R
\name{zoomPlot}
\alias{zoomPlot}
\title{Zoomplot}
\usage{
zoomPlot(m, varnames = NULL, mn = NULL, mx = NULL, cll = NULL,
  hoof = "", ...)
}
\arguments{
\item{m}{Data frame}

\item{varnames}{Character vector. Variable names}

\item{mn}{Date object. Refers to the minimum date}

\item{mx}{Date object. Refers to the maximum date}

\item{cll}{Character vector. Guide color label}

\item{hoof}{Character vector. The header of the plot. Does not have a header
by default}

\item{...}{Arguments to be passed to/from calling function}
}
\description{
Generic Function
}

