% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomVuur.R
\name{zoomVuur}
\alias{zoomVuur}
\title{Zoom Vuur}
\usage{
zoomVuur(x, mn = NULL, mx = NULL, hoof = "",
  timeZone = "Africa/Johannesburg", useOnlyNumeric = TRUE)
}
\arguments{
\item{x}{Data frame. Contains fire data}

\item{mn}{Date object. Refers to the minimum date}

\item{mx}{Date object. Refers to the maximum date}

\item{hoof}{Character vector. The header of the plot. Does not have a header
by default}

\item{timeZone}{Character vector. The time zone under consideration}

\item{useOnlyNumeric}{Logical. Plots only numeric columns when TRUE}
}
\description{
Creates a line and point plot of data and allows the user to manipulate the data
by being able to zoom in or out of the date axis
}
\details{
Use in format x = vuurlys[[2]] or
structure(list(B = c(15, 15, 14.75, 14.5, 14.5, 14.5), K = c(15, 15, 14.75, 14.5, 14.5, 14.5), L = c(14.5, 14.5, 14.5, 14.5, 14.5,  14.5), S = c(14.75, 14.5, 14.5, 14.5, 14.5, 14.5), sdif = c(-0.0833333333333339,
-0.333333333333334, -0.166666666666666, 0, 0, 0), vuur = structure(c(1L, 1L, 1L, 1L, 1L, 1L), .Label = c("no", "fire"), class = "factor"), date = c("2013-07-04 08:00:00", "2013-07-04 08:15:00", "2013-07-04 08:30:00",
"2013-07-04 08:45:00", "2013-07-04 09:00:00", "2013-07-04 09:15:00")), .Names = c("B", "K", "L", "S", "sdif", "vuur", "date"), row.names = c("2013-07-04 08:00:00", "2013-07-04 08:15:00",
"2013-07-04 08:30:00", "2013-07-04 08:45:00", "2013-07-04 09:00:00",  "2013-07-04 09:15:00"), class = "data.frame")
}

