% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maak.vuur.R
\name{pl.pca}
\alias{pl.pca}
\title{Experimental Plotting Function}
\usage{
pl.pca(x = dpca$ind$coord[, 1], y = ddpca$ind$coord[, 2], tp = "p",
  sdif = "sdif", sdifval = 3, pcadf = cluscont, cl, clII, mn, nm, nmII,
  envr = zz, qplot = TRUE, verbose = TRUE, naam = "naam")
}
\description{
Function is used to plot data output from the PCA
}

